optimize service update mechanism
